delay: Add delay module (#23)
This implements the embedded-hal 1.0 `DelayNs` trait. The cortex-m crate
has essentially the same logic, but the published version is based on
embedded-hal 0.2. This can be removed (repurposed for use with timers?)
when a new version of the cortex-m crate is published.

The blinky example has been updated to use the delay module.