gpdma: add GPDMA driver
This adds support for one-shot DMA transfers for all supported transfer
types with linear addressing. It does not yet support linked list buffer
loading, or support 2D addressing.

Memory-to-memory transfers were quite thoroughly tested by the provided
example, memory-to-peripheral and peripheral-to-memory transfers were
tested with the SPI peripheral (working on a test branch).
Peripheral-to-peripheral transfers are assumed to work given their
similarity to memory-to-peripheral/peripheral-to-memory transfers,
but will be properly tested at a later stage.

The driver includes helper structs for peripheral transfers in the
gpdma::periph module which handle the common operations for setting up
one-directional and full-duplex transfers using one, or two channels,
respectively.