i2c: Add I2C Controller implementation (#27)
This adds the I2C Controller driver implementation for the STM32H5. It
implements a blocking, and non-blocking API, as well as the embedded-hal
1.0 traits.

It borrows a lot from the STM32H7 HAL (particularly the TIMINGR
configuration, which is a bit inscrutable in the reference manual), but
it borrows the `Instance` concept from the STM32F4 project to minimize
the need for macros.

It uses the updated terminology of Controller/Target from the latest
version of the I2C spec (Rev. 7.0)