Fix OSPEEDR and PUPDR indexing
The field width of 2 was being applied twice in these two registers,
because the PAC already accounts for it, so there is no need to apply
it externally.

Thankfully the PAC triggers and out-of-bounds access to an array when
this happens in debug builds, so it can be easily caught.