spi: Add SPI master implementation (#53)
This adds a master mode implementation for the SPI peripheral. 

The driver is split such that the core logic is implemented in spi.rs,
and the rest is contained within a submodule:
- User configuration is split off into the config module
- The transaction module encapsulates the different transaction types
(read, write, transfer, transfer inplace). Buffer management is
abstracted into the transaction type to enable non-blocking usage the
peripheral.
- The public facing APIs are mostly implemented in the hal module
(embedded-hal trait implementations), and the nonblocking module, which
exposes a NonBlocking trait that a user needs to explicitly use in order
to use the non-blocking APIs (expected to be the less used case, this
avoids polluting the public API too much).