i2c: Add target implementation (#45)
This extends the I2C driver to support target operation. It adds an
I2cTarget type to represent the I2C operation as the target.

Target operation can be operated in automatic ACKing mode, letting the
hardware manage clock stretching and ACKing automatically, or it can be
operated in manual ACK control mode, which lets the user of the driver
determine when to ACK the end of a transaction. Manual ACK control
allows for clock stretching while data is being processed or retrieved
before ACK'ing the final byte of a transfer and initiating either the
end of a transfer or a repeat start to read the processed/retrieved
data. This implementation does not currently allow for ACKing individual
or a subset of bytes manually, but extending the interface to do so can
build upon what is implemented so far.

This also adds role switching to the driver, enabling both controller
and target implementation through choosing the appropriate
initialization functions.

---------

Co-authored-by: Albin Hedman <albin9604@gmail.com>