DMA Futures (#65)
This implements futures for GPDMA transfers by adding a wrapper for
`DmaTransfer`, `DmaTransferFuture`, implementing Future for
`DmaTransferFuture` in the `gpdma::future` module, which is created when
a `DmaTransfer` is awaited via the `IntoFuture` trait. Under the hood,
the implementation uses an `AtomicWaker` from the
[`futures-util`](https://docs.rs/futures-util/latest/futures_util/)
crate and channel interrupts to drive the future. Because the wakers
need to be defined statically in order to be accessible from the channel
interrupts, and to prevent accidental colliding implementations of
interrupt handlers for the DMA channels, the futures implementation is
hidden behind the `gpdma-future` feature flag.